PSIMU starting to work on linux :)

+++ trunk/code/linux/IRTSLinux/Patient.h
@@ -22,7 +22,7 @@ class Patient
 	SignalIterator*	getECG();
 	AnnotIterator*	getAnn();
 
-	int                  samplefreq;
+	double          samplefreq;
 
  private:
 	DbPatient* _patient;
